working refactor
